Move quantized fused ops to cpu/op
