+++ src/ngraph/op/op_tbl.hpp
@@ -121,12 +121,7 @@ NGRAPH_OP(Power, ngraph::op)
 NGRAPH_OP(Product, ngraph::op)
 NGRAPH_OP(Quantize, ngraph::op)
 NGRAPH_OP(QuantizedConvolution, ngraph::op)
-NGRAPH_OP(QuantizedConvolutionBias, ngraph::op)
-NGRAPH_OP(QuantizedConvolutionBiasAdd, ngraph::op)
-NGRAPH_OP(QuantizedConvolutionBiasSignedAdd, ngraph::op)
-NGRAPH_OP(QuantizedConvolutionRelu, ngraph::op)
 NGRAPH_OP(QuantizedDot, ngraph::op)
-NGRAPH_OP(QuantizedDotBias, ngraph::op)
 NGRAPH_OP(Recv, ngraph::op)
 NGRAPH_OP(Range, ngraph::op)
 NGRAPH_OP(Relu, ngraph::op)
