+++ src/ngraph/runtime/cpu/op/quantized_conv_relu.cpp
@@ -14,7 +14,7 @@
 // limitations under the License.
 //*****************************************************************************
 
-#include "ngraph/op/experimental/quantized_conv_relu.hpp"
+#include "quantized_conv_relu.hpp"
 #include "ngraph/op/convolution.hpp"
 #include "ngraph/op/get_output_element.hpp"
