+++ src/ngraph/runtime/cpu/op/quantized_conv_relu.hpp
@@ -18,6 +18,7 @@
 
 #include "ngraph/coordinate_diff.hpp"
 #include "ngraph/op/op.hpp"
+#include "ngraph/runtime/cpu/cpu_backend_visibility.h"
 
 namespace ngraph
 {
@@ -27,7 +28,7 @@ namespace ngraph
         class QuantizedConvolutionRelu : public Op
         {
         public:
-            NGRAPH_API
+            CPU_BACKEND_API
             static const std::string type_name;
             const std::string& description() const override { return type_name; }
             QuantizedConvolutionRelu() = default;
