+++ src/ngraph/runtime/cpu/op/quantized_dot_bias.cpp
