+++ src/ngraph/serializer.cpp
@@ -59,9 +59,6 @@
 #include "ngraph/op/experimental/dyn_reshape.hpp"
 #include "ngraph/op/experimental/dyn_slice.hpp"
 #include "ngraph/op/experimental/generate_mask.hpp"
-#include "ngraph/op/experimental/quantized_conv_bias.hpp"
-#include "ngraph/op/experimental/quantized_conv_relu.hpp"
-#include "ngraph/op/experimental/quantized_dot_bias.hpp"
 #include "ngraph/op/experimental/range.hpp"
 #include "ngraph/op/experimental/shape_of.hpp"
 #include "ngraph/op/experimental/tile.hpp"
@@ -1601,14 +1598,6 @@ shared_ptr<Node> JSONDeserializer::deserialize_node(json node_js)
             node = make_shared<op::Quantize>(args[0], args[1], args[2], type, axes, round_mode);
             break;
         }
-        case OP_TYPEID::QuantizedConvolutionBias: { break;
-        }
-        case OP_TYPEID::QuantizedConvolutionBiasAdd: { break;
-        }
-        case OP_TYPEID::QuantizedConvolutionBiasSignedAdd: { break;
-        }
-        case OP_TYPEID::QuantizedConvolutionRelu: { break;
-        }
         case OP_TYPEID::QuantizedConvolution:
         {
             auto window_movement_strides =
@@ -1642,8 +1631,6 @@ shared_ptr<Node> JSONDeserializer::deserialize_node(json node_js)
                 output_axes);
 
             break;
-        }
-        case OP_TYPEID::QuantizedDotBias: { break;
         }
         case OP_TYPEID::QuantizedDot:
         {
@@ -2695,14 +2682,6 @@ json JSONSerializer::serialize_node(const Node& n)
         node["round_mode"] = tmp->get_round_mode();
         break;
     }
-    case OP_TYPEID::QuantizedConvolutionBias: { break;
-    }
-    case OP_TYPEID::QuantizedConvolutionBiasAdd: { break;
-    }
-    case OP_TYPEID::QuantizedConvolutionBiasSignedAdd: { break;
-    }
-    case OP_TYPEID::QuantizedConvolutionRelu: { break;
-    }
     case OP_TYPEID::QuantizedConvolution:
     {
         auto tmp = dynamic_cast<const op::QuantizedConvolution*>(&n);
@@ -2716,8 +2695,6 @@ json JSONSerializer::serialize_node(const Node& n)
         node["filter_axes"] = tmp->get_filter_axes();
         node["output_axes"] = tmp->get_output_axes();
         break;
-    }
-    case OP_TYPEID::QuantizedDotBias: { break;
     }
     case OP_TYPEID::QuantizedDot:
     {
